Create trade_recap.m
This is a Matlab script: extract daily trade recap data and process the data, then write them into a 'csv' interface, waiting for manual approvement. If the trade recap files from broker are approved, change the approved flag to '1', if not, change it to '2'. Another batch file will monitor the 'csv' interface to do later process.